fix: fix for pgstore test
issue with using a map to iterate through values and expecting order to
be retained. this resulted in intermittent situations where the values
returned were not in an expected order. the fix here to is to add an
ordered slice of keys along side the map to retain order of the pgtx
cache. when the batch method runs, will persist according to this order.